refactor: split out `UPOS` from `DictWordMetadata`
